requirements.txt
# Core AI Frameworks
openai>=1.30.0
langgraph>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0

# Search Engines
tavily-python>=0.3.0
duckduckgo-search>=5.0.0
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
aiohttp>=3.9.0
asyncio
uuid
datetime
re
json
os

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Azure OpenAI
azure-openai>=1.0.0

# Logging
structlog>=23.0.0

# HTTP Client for API calls
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration Management
pyyaml>=6.0.0

# Progress Bars
rich>=13.0.0

# File Handling
pathlib2>=2.3.0

# Date/Time Handling
python-dateutil>=2.8.0

# Optional: Jupyter Support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Development Tools
pre-commit>=3.0.0
mypy>=1.0.0
isort>=5.0.0